physics refactor, added GraphicsDevice, winwos fixes

+++ Window/WindowsWindow.cpp
@@ -32,8 +32,8 @@ LRESULT Core::WindowsWindow::window_procedure(HWND hwnd, UINT msg, WPARAM wParam
 	{
 		RECT r = *(RECT*)lParam;
 		
-		const auto window_width = r.right - r.left;
-		const auto window_height = r.bottom - r.top;
+		const auto window_width  = (UINT)( r.right - r.left );
+		const auto window_height = (UINT)( r.bottom - r.top );
 		
 		if (window_width > window->max_width)
 			((RECT*)lParam)->right = r.left + window->max_width;
